Add avatar job handling
