+++ src/GRA.Controllers/MissionControl/AvatarsController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -29,12 +30,14 @@ namespace GRA.Controllers.MissionControl
         public AvatarsController(ILogger<AvatarsController> logger,
             ServiceFacade.Controller context,
             AvatarService avatarService,
+            JobService jobService,
             IHostingEnvironment hostingEnvironment)
             : base(context)
         {
-            _logger = Require.IsNotNull(logger, nameof(logger));
-            _avatarService = Require.IsNotNull(avatarService, nameof(avatarService));
-            _hostingEnvironment = Require.IsNotNull(hostingEnvironment, nameof(hostingEnvironment));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _avatarService = avatarService ?? throw new ArgumentNullException(nameof(avatarService));
+            _jobService = jobService ?? throw new ArgumentNullException(nameof(jobService));
+            _hostingEnvironment = hostingEnvironment ?? throw new ArgumentNullException(nameof(hostingEnvironment));
             PageTitle = "Avatars";
         }
 
@@ -65,8 +68,6 @@ namespace GRA.Controllers.MissionControl
                 return RedirectToAction(nameof(Index));
             }
 
-            int siteId = GetCurrentSiteId();
-
             string assetPath = Path.Combine(
                 Directory.GetParent(_hostingEnvironment.WebRootPath).FullName, "assets");
 
@@ -85,7 +86,12 @@ namespace GRA.Controllers.MissionControl
 
             var jobToken = await _jobService.CreateJobAsync(new Job
             {
-                JobType = JobType.AvatarImport
+                JobType = JobType.AvatarImport,
+                SerializedParameters = JsonConvert
+                    .SerializeObject(new JobDetailsAvatarImport
+                    {
+                        AssetPath = assetPath
+                    })
             });
 
             return View("Job", new ViewModel.MissionControl.Shared.JobViewModel
@@ -97,158 +103,6 @@ namespace GRA.Controllers.MissionControl
                 SuccessUrl = Url.Action(nameof(Index)),
                 Title = "Importing avatars..."
             });
-
-            IEnumerable<AvatarLayer> avatarList;
-            var jsonPath = Path.Combine(assetPath, "default avatars.json");
-            using (StreamReader file = System.IO.File.OpenText(jsonPath))
-            {
-                var jsonString = await file.ReadToEndAsync();
-                avatarList = JsonConvert.DeserializeObject<IEnumerable<AvatarLayer>>(jsonString);
-            }
-
-            _logger.LogInformation($"Found {avatarList.Count()} AvatarLayer objects in avatar file");
-
-            var time = _dateTimeProvider.Now;
-            int totalFilesCopied = 0;
-            var userId = GetId(ClaimType.UserId);
-
-            var backgroundRoot = Path.Combine($"site{siteId}", "avatarbackgrounds");
-            var backgroundPath = _pathResolver.ResolveContentFilePath(backgroundRoot);
-            if (!Directory.Exists(backgroundPath))
-            {
-                Directory.CreateDirectory(backgroundPath);
-            }
-            System.IO.File.Copy(Path.Combine(assetPath, "background.png"),
-                Path.Combine(backgroundPath, "background.png"));
-            totalFilesCopied++;
-
-            foreach (var layer in avatarList)
-            {
-                int layerFilesCopied = 0;
-
-                var colors = layer.AvatarColors;
-                var items = layer.AvatarItems;
-                layer.AvatarColors = null;
-                layer.AvatarItems = null;
-
-                var addedLayer = await _avatarService.AddLayerAsync(layer);
-
-                var layerAssetPath = Path.Combine(assetPath, addedLayer.Name);
-                var destinationRoot = Path.Combine($"site{siteId}", "avatars", $"layer{addedLayer.Id}");
-                var destinationPath = _pathResolver.ResolveContentFilePath(destinationRoot);
-                if (!Directory.Exists(destinationPath))
-                {
-                    Directory.CreateDirectory(destinationPath);
-                }
-
-                addedLayer.Icon = Path.Combine(destinationRoot, "icon.png");
-                System.IO.File.Copy(Path.Combine(layerAssetPath, "icon.png"),
-                    Path.Combine(destinationPath, "icon.png"));
-
-                await _avatarService.UpdateLayerAsync(addedLayer);
-
-                if (colors != null)
-                {
-                    foreach (var color in colors)
-                    {
-                        color.AvatarLayerId = addedLayer.Id;
-                        color.CreatedAt = time;
-                        color.CreatedBy = userId;
-                    }
-                    await _avatarService.AddColorListAsync(colors);
-                    colors = await _avatarService.GetColorsByLayerAsync(addedLayer.Id);
-                }
-                foreach (var item in items)
-                {
-                    item.AvatarLayerId = addedLayer.Id;
-                    item.CreatedAt = time;
-                    item.CreatedBy = userId;
-                }
-                await _avatarService.AddItemListAsync(items);
-                items = await _avatarService.GetItemsByLayerAsync(addedLayer.Id);
-
-                var elementList = new List<AvatarElement>();
-                _logger.LogInformation($"Processing {items.Count} items in {addedLayer.Name}...");
-
-                foreach (var item in items)
-                {
-                    var itemAssetPath = Path.Combine(layerAssetPath, item.Name);
-                    var itemRoot = Path.Combine(destinationRoot, $"item{item.Id}");
-                    var itemPath = Path.Combine(destinationPath, $"item{item.Id}");
-                    if (!Directory.Exists(itemPath))
-                    {
-                        Directory.CreateDirectory(itemPath);
-                    }
-                    item.Thumbnail = Path.Combine(itemRoot, "thumbnail.jpg");
-                    System.IO.File.Copy(Path.Combine(itemAssetPath, "thumbnail.jpg"),
-                        Path.Combine(itemPath, "thumbnail.jpg"));
-                    if (colors != null)
-                    {
-                        foreach (var color in colors)
-                        {
-                            var element = new AvatarElement()
-                            {
-                                AvatarItemId = item.Id,
-                                AvatarColorId = color.Id,
-                                Filename = Path.Combine(itemRoot, $"item_{color.Id}.png")
-                            };
-                            elementList.Add(element);
-                            System.IO.File.Copy(
-                                Path.Combine(itemAssetPath, $"{color.Color}.png"),
-                                Path.Combine(itemPath, $"item_{color.Id}.png"));
-                            layerFilesCopied++;
-                        }
-                    }
-                    else
-                    {
-                        var element = new AvatarElement()
-                        {
-                            AvatarItemId = item.Id,
-                            Filename = Path.Combine(itemRoot, "item.png")
-                        };
-                        elementList.Add(element);
-                        System.IO.File.Copy(Path.Combine(itemAssetPath, "item.png"),
-                            Path.Combine(itemPath, "item.png"));
-                        layerFilesCopied++;
-                    }
-                }
-
-                await _avatarService.UpdateItemListAsync(items);
-                await _avatarService.AddElementListAsync(elementList);
-                totalFilesCopied += layerFilesCopied;
-                _logger.LogInformation($"Copied {layerFilesCopied} items for {layer.Name}");
-            }
-            _logger.LogInformation($"Copied {totalFilesCopied} items for all layers.");
-
-            var bundleJsonPath = Path.Combine(assetPath, "default bundles.json");
-            if (System.IO.File.Exists(bundleJsonPath))
-            {
-                IEnumerable<AvatarBundle> bundleList;
-                using (StreamReader file = System.IO.File.OpenText(bundleJsonPath))
-                {
-                    var jsonString = await file.ReadToEndAsync();
-                    bundleList = JsonConvert.DeserializeObject<IEnumerable<AvatarBundle>>(jsonString);
-                }
-
-                foreach (var bundle in bundleList)
-                {
-                    _logger.LogInformation($"Processing bundle {bundle.Name}...");
-                    var items = new List<int>();
-                    foreach (var bundleItem in bundle.AvatarItems)
-                    {
-                        var item = await _avatarService.GetItemByLayerPositionSortOrderAsync(
-                            bundleItem.AvatarLayerPosition, bundleItem.SortOrder);
-                        items.Add(item.Id);
-                    }
-                    bundle.AvatarItems = null;
-                    await _avatarService.AddBundleAsync(bundle, items);
-                }
-            }
-
-            string loaded = $"Default avatars added in {5} seconds.";
-            _logger.LogInformation(loaded);
-            ShowAlertSuccess(loaded);
-            return RedirectToAction(nameof(Index));
         }
 
         public async Task<IActionResult> Layer(int id,
